requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
codecov==2.1.12
colorama==0.4.5
coverage==6.4.4
fastapi==0.85.0
fastapi-versioning==0.10.0
fpl==0.6.34
frozenlist==1.3.1
h11==0.14.0
idna==3.4
iniconfig==1.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
PTable==0.9.2
py==1.11.0
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.1.3
pytest-aiohttp==1.0.4
pytest-asyncio==0.19.0
pytest-cov==4.0.0
pytest-mock==3.9.0
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
yarl==1.8.1
